Update state manager debug view based on design meeting feedback
Issue #19109

* Moved to StateManager
* Implemented as ToDebugString on InternalEntityEntry
* Used for ToString on InternalEntityEntry
* Also for ToString on external EntityEntry
* Also added DebugView on InternalEntityEntry to access long form for single entry
