+++ src/EFCore/ChangeTracking/ChangeTracker.cs
@@ -10,7 +10,6 @@ using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
 using Microsoft.EntityFrameworkCore.Infrastructure;
-using Microsoft.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Utilities;
 
@@ -396,18 +395,6 @@ namespace Microsoft.EntityFrameworkCore.ChangeTracking
 
             return default;
         }
-
-        /// <summary>
-        ///     This is an internal API that supports the Entity Framework Core infrastructure and not subject to
-        ///     the same compatibility standards as public APIs. It may be changed or removed without notice in
-        ///     any release. You should only use it directly in your code with extreme caution and knowing that
-        ///     doing so can result in application failures when updating to a new Entity Framework Core release.
-        /// </summary>
-        [EntityFrameworkInternal]
-        public virtual DebugView DebugView
-            => new DebugView(
-                () => StateManager.ToDebugString(StateManagerDebugStringOptions.ShortDefault),
-                () => StateManager.ToDebugString(StateManagerDebugStringOptions.LongDefault));
 
         #region Hidden System.Object members
