+++ src/EFCore/Update/UpdateEntryExtensions.cs
@@ -28,7 +28,16 @@ namespace Microsoft.EntityFrameworkCore.Update
             [NotNull] IEnumerable<IPropertyBase> properties)
             => "{"
                 + string.Join(
-                    ", ", properties.Select(p => p.Name + ": " + Convert.ToString(entry.GetCurrentValue(p), CultureInfo.InvariantCulture)))
+                    ", ", properties.Select(
+                        p =>
+                        {
+                            var currentValue = entry.GetCurrentValue(p);
+                            return p.Name
+                                + ": "
+                                + (currentValue == null
+                                    ? "<null>"
+                                    : Convert.ToString(currentValue, CultureInfo.InvariantCulture));
+                        }))
                 + "}";
 
         /// <summary>
@@ -44,7 +53,16 @@ namespace Microsoft.EntityFrameworkCore.Update
             [NotNull] IEnumerable<IPropertyBase> properties)
             => "{"
                 + string.Join(
-                    ", ", properties.Select(p => p.Name + ": " + Convert.ToString(entry.GetOriginalValue(p), CultureInfo.InvariantCulture)))
+                    ", ", properties.Select(
+                        p =>
+                        {
+                            var originalValue = entry.GetOriginalValue(p);
+                            return p.Name
+                                + ": "
+                                + (originalValue == null
+                                    ? "<null>"
+                                    : Convert.ToString(originalValue, CultureInfo.InvariantCulture));
+                        }))
                 + "}";
     }
 }
